Split ShowSection into multiple files
