+++ composeApp/src/main/kotlin/io/github/couchtracker/ui/screens/main/ShowSectionViewModel.kt
@@ -28,6 +28,7 @@ import io.github.couchtracker.tmdb.toBaseShow
 import io.github.couchtracker.ui.components.ShowPortraitModel
 import io.github.couchtracker.ui.components.UpNextListItemModel
 import io.github.couchtracker.ui.components.toShowPortraitModels
+import io.github.couchtracker.ui.screens.main.show.ShowExploreTabState
 import io.github.couchtracker.utils.Loadable
 import io.github.couchtracker.utils.Result
 import io.github.couchtracker.utils.collectAsLoadable
